Foreign.OCaml

Description

Documentation

caml_startup :: [String] -> IO ()

Startup for the OCaml runtime: call this function before using any other function in this module.

class Marshal t where

Marshal provides functions for transforming a Haskell representation of data into OCaml and viceversa. Default implementations are given for all basic types: Unit, Bool, Int, Float, Double, String; and dependent declarations allow to construct Marshal implementations for lists, tuples and applicative types. Note that for applicative types the marshal function is not implemented, so it is not (yet) possible to pass higher order values.

Methods

marshal :: t -> Value

marshal takes a Haskell value of type t and produces an opaque OCaml value.

unmarshal :: Value -> t

unmarshal takes an OCaml value and returns its deserialization into a Haskell data type

deriveMarshalInstance :: Name -> Q [Dec]

Automatically generate an instance of Marshal given a datatype. If you have an OCaml data type with no type variables and multiple constructors:

type t = C1 of int * string | C2 of t * t | C3

Simply produce a corresponding Haskell data type, then call "deriveMarshalInstance" to obtain two way conversion between the Haskell and OCaml representations:

data T = C1 Int String | C2 T T | C3
$(deriveMarshalInstance ''T)

type Value = CLong

Opaque datatype that represents an OCaml object.

register_closure :: Marshal a => String -> a

register_closure returns a closure of type a for any OCaml name that has been registered with:

let _ = Callback.register "f" f

If f is a function of OCaml type `int -> int` then it can be registered with:

f = register_closure "f" :: Int -> Int

register_closure works with type variables as well, as long as all its instantiations are also instance of Marshal. For example the function linearize that takes a tree and returns a list, with parameter type a:

type 'a tree = Leaf of 'a | Node of 'a tree * 'a tree;;
let rec linearize t = match t with
  | Leaf(x)   -> [x]
  | Node(l,r) -> let ll = linearize l in
                 let lr = linearize r in
                 ll @ lr;;
let _ = Callback.register "linearize" linearize

can be registered as such:

linearize = register_closure "linearize" :: Tree a -> [a]

provided that an implementation of "Tree a" and Marshal for both "Tree a" and "a" exist.

block_constructor :: Tag -> [Value] -> Value

Produces a block constructor, that is, a constructor with arguments. The integer pararmeter is the index of the block constructor, in order of appearance on the declaration of the datatype, counting from 0 and not considering constant constructors.

constant_constructor :: Int -> Value

Produces a constant constructor, that is, a constructor that has no parameters. The integer parameter is the index of the constant constructor, in order of appearance on the declaration of the datatype, counting from 0 and not considering block constructors.

is_block :: Value -> Bool

True if v is a block constructor

is_value :: Value -> Bool

True if v is an atomic value

get_const :: Value -> Int

Returns the index for a constant constructor, inverse of constant_constructor.

get_field :: Value -> Int -> Value

Returns a field in position offset from a block constructor.

get_tag :: Value -> CUChar

Returns the idnex for a block constructor, more or less an inverse of block_constructor, to be used together with get_field